requirements.txt
alembic==1.9.0
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==21.2.0
bcc==0.18.0
bcrypt==4.0.1
billiard==3.6.4.0
blinker==1.4
Brlapi==0.8.3
celery==5.2.7
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
databases==0.6.2
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
dnspython==2.2.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.18.0
email-validator==1.3.0
et-xmlfile==1.1.0
fastapi==0.88.0
Flask==2.2.2
Flask-Celery==2.4.3
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==2.0.1
gyp==0.1
h11==0.14.0
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jose==1.0.0
jsonschema==3.2.0
jwt==1.3.1
keyring==23.5.0
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.2.4
MarkupSafe==2.1.1
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
numpy==1.24.1
oauthlib==3.2.0
openpyxl==3.0.10
pandas==1.5.2
passlib==1.7.4
pop-transition==1.1.2
prompt-toolkit==3.0.37
protobuf==3.12.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.2
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
python-gnupg==0.4.8
python-jose==3.3.0
python-xlib==0.29
pytz==2022.7
pyxdg==0.27
PyYAML==5.4.1
redis==4.5.1
repolib==2.2.0
repoman==1.4.0
requests==2.25.1
rsa==4.9
SecretStorage==3.3.1
secure-smtplib==0.1.1
sessioninstaller==0.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy-mptt==0.2.5
starlette==0.22.0
systemd-python==234
texttable==1.6.4
typing_extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
uvicorn==0.20.0
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.6
websocket-client==1.2.3
Werkzeug==2.2.2
WTForms==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
